+++ packages/core/lib/search/metaheuristics/evolutionary/EvolutionaryAlgorithm.ts
@@ -26,6 +26,7 @@ import { CONFIG } from "../../../Configuration";
 import { TerminationManager } from "../../termination/TerminationManager";
 import { Encoding } from "../../Encoding";
 import { tournamentSelection } from "../../operators/selection/TournamentSelection";
+import { EventManager } from "../../../event/EventManager";
 
 /**
  * Base class for Evolutionary Algorithms (EA).
@@ -58,6 +59,7 @@ export abstract class EvolutionaryAlgorithm<
   /**
    * Constructor.
    *
+   * @param eventManager The event manager
    * @param objectiveManager The objective manager used by the specific algorithm
    * @param encodingSampler The encoding sampler used by the specific algorithm
    * @param crossover The crossover operator to apply
@@ -65,11 +67,12 @@ export abstract class EvolutionaryAlgorithm<
    * @protected
    */
   protected constructor(
+    eventManager: EventManager<T>,
     objectiveManager: ObjectiveManager<T>,
     encodingSampler: EncodingSampler<T>,
     crossover: Crossover<T>
   ) {
-    super(objectiveManager);
+    super(eventManager, objectiveManager);
     this._encodingSampler = encodingSampler;
     this._population = [];
     this._populationSize = CONFIG.populationSize;